Reduce thrown exceptions when creating DisposableSaveableStateRegistry
In cases like ComposeViews in RecyclerView, where no View ID is set on
several sibling ComposeViews, DisposableSaveableStateRegistry creation
uses an exception to figure out that it can't register the View as
saveable. This change first checks to see if the View is already in the
registry before registering so that it doesn't need to throw an exception.

Test: existing tests
Change-Id: Ib1a5fb5a9326151204a52f0c02acdae8abbce6fe