Make shared element map observable for transition activeness observation
This change makes the shared element map observable, so that when
new shared elements are added to the map the snapshot observation
gets updated to observe the new shared elements.

Previously, the activeness is explicitly updated when any shared element
entry is added/removed. This covered the common use case for
shared elements where the outgoing shared elements are removed
from the tree once the animation is finished. But it isn't
sufficient to handle cases where
sharedElementWithCallerManagedVisibility is used and therefore
there is no event of shared element removal, causing the
transition activeness to not be updated.

Bug: 450752872
Test: new test added
Change-Id: I5ad9969c004412cde8f5604740a36b9086e66059