Ignore consumed down events in TextFields
The common selection gesture detector for both SelectionContainer and TextFields do not ignore the consumed down events. For TextFields this presents a problem since material buttons that are placed in the decorator area also pass on their down events to the TextFields. These consumed events shouldn't trigger any TextField interaction.

Test: ./gradlew :compose:foundation:foundation:cAT
Bug: 456608928
Change-Id: Id9d2033369e7b488c0b1989523f8362f921ea98c