Only recalculate position in root when something changed
Now, that we have a good test coverage and are confident that onCoordinatorRectChanged() function is called whenever there is a bounds change, we can do the optimization to skip all the rectmanager invalidations until a change in position/size is detected. It will be especially beneficial for the layer transformed subtrees, as the calculations there are more expensive. We still only recalculate the children bounds once when all the changes for the modifiers on the parent are finalized.

Test: existing tests
Change-Id: Ia8eebb82ec92d9ca99c38fad01cec5b503d994f1