Switch some issue suppressions from property to field
Previously, annotations on a property would end up on the corresponding field in UAST, but that does not always make sense if the use-site target of the annotation was not for the field.

This moves some `@Suppress` annotations to the field use-site target to suppress issues reported on the fields.

Bug: 450916834
Test: ./gradlew checkApi
Relnote: N/A
Change-Id: Icee9f402a163b58eadbbbe80cbbc472dd9875e39