Check if entity is disposed.
Add the ability to check if an entity is disposed before executing actions on it. This modifies the ActionQueue implementation to allow an "isValid" method that will be checked before performing actions on the value to ensure that the value is in a valid state prior to executing actions in the queue.

This also changes the signature of the "executeWhenAvailable" method to take the value as a parameter instead of as the scope for the function. This should make it more clear which actions in the block are applied to the value object versus a local function.

Bug: 440426914
Test: Added unit tests and unit tests pass.
Change-Id: I0c1f3b4cc05b884197f0dac07af4390cb2429224