Add benchmarks for NavHost in Compose
This commit introduces a new benchmark test, `NavHostBenchmark`, to measure the performance of the `NavHost` composable.

The benchmark measures the time to first pixel for a `NavHost` with one, two, and ten composable destinations.

To support this, the `navigation-benchmark` module has been configured to build with Compose by enabling the `AndroidXComposePlugin` and adding the necessary Compose dependencies. The `compileSdk` has also been updated to 35.

Additionally, this change removes workaround dependencies for `-ktx` artifacts that are no longer needed.

Test: Added NavHostBenchmarks
Change-Id: Ie84c927823c693d26e4790dee8522c553ce1d0f1