Track accessors of PublishedApi properties
Internal APIs annotated with @PublishedApi need to be tracked for binary compatibility because they can be used from (non-internal) inline functions, which means they can end up inlined in bytecode for external clients which use the inline functions.

Previously, the accessor methods of @PublishedApi properties weren't being tracked, but they are how the property usages will be inlined in bytecode, so they should be.

Test: ./gradlew checkApi
Bug: 450844478
Relnote: N/A
Change-Id: I9a2c96b077fef0770599fe9b3947818931fb0d42