Fix SceneCoreEntity layout to respect constraints.
The SceneCoreEntity layout logic was only ensuring that the lower bound of the constraints were enforced but not the upper bound of the constraints.

This fixes the case where if a state change updates the logic of the sizeAdapter then it should retrigger composition and layout.

This changes SubspaceLayoutNode to only consider its layout size instead of querying the CoreEntity for its size. It should be up to the measure policies to determine how the entities' sizes are used to influence layout. The SubspaceLayoutNode should not get its size from the entity.

Bug: 458483058
Test: Added unit tests
Fixes: 458483058
Change-Id: I11bb838de7fd357f09febfa63a743bf78aa78a2e
Relnote: Fixed a layout bug in SceneCoreEntity. It should now properly respect its constraints.