Remove ignoreRemeasureRequests() wrapping PausedComposition.resume call in SubcomposeLayout
It was just a copy paste from the regular composition code, but really we don't need this, as resuming pausable composition is never changing the LayoutNode's hierarchy, it is done separately as part of apply, where we have ignoreRemeasureRequests, and it is still needed.
Removing this helps to fix potential threading issues if one try to call resume from another thread.

Test: existing tests
Change-Id: I8f385388235d72d21febf500ea697aa0feda10ab