Refactor Anchorable.kt
The Anchorable node should not need to add a move event listener in
order to function properly, as the movement is controlled by the system.
The pose set for the entity during composition can be retrieved from the
CoreEntity associated with the Entity the MovableComponent is attached
to.

Adds a getter for the latest runtime movable component created by a
FakeSceneRuntime for the purpose of injecting move events in unit tests.

Bug: 473598187
Test: Tested SpatialCompose activity on Moohan device

Change-Id: I6ce8869f8de1369f061b9a5ddfe4f0b303a27693