Add two scripts to help query information about released artifacts
- get_release_information.py

This script will take a maven coordinate of an androidx library such as
androidx.collection:collection:1.5.0 and returns two links, one to the
build page and another to the +source tree at the state of the release.

./development/get_release_information.py androidx.browser:browser:1.9.0
./development/get_release_information.py androidx.activity:activity:1.11.0
./development/get_release_information.py androidx.collection:collection:1.5.0

- get_release_diff.py

This script will take a mave coordinate of an androidx library such as
androidx.collection:collection and two version such as 1.5.0 and
1.6.0-alpha01 and return two links, one to the source diff between the
two releases and another to the list of commits between the two
releases.

./development/get_release_diff.py androidx.compose.runtime:runtime 1.8.0 1.9.0
./development/get_release_diff.py androidx.collection:collection 1.5.0 1.6.0-alpha01

Test: running the scripts
Change-Id: I170e00126da64656741f2afed719915128900a2a