Align onGloballyPositioned and RectManager callbacks dispatching timings
We want those two dispatching mechanisms to be interchangeable in order to be able to migrate more usages into the most efficient one. Inconsistency in those timings caused tests to fail in g3 after we migrated AwaitFirstLayoutModifier to registerOnLayoutRectChanged in aosp/3761253. The same issue was affecting our internal tests when I tried to locally update onGloballyPositioned to use this new dispatching.

I wasn't able to write a clear test specifically testing what was causing an issue in g3, as there is a combination of many factors (UnconfinedTestDispatcher, which we don't use anymore; nested lazy layouts both are doing scrolling from LaunchedEffect; AndroidViews in the subtree), but I verified that those g3 tests are passing now, and our existing tests are passing when we use new dispatch for onGloballyPositioned.

Test: run affected g3 tests
Fixes: 458833945
Fixes: 458832089

Change-Id: If549216232420f67c2a86a1895e946c8ce0c185c