Update signature files
Updating signature files for changes related to tracking exhaustivity
for sealed classes.

We want to track whether or not a sealed class is "exhaustive"
(instances of it can be exhaustively matched) to more easily determine
if future changes will cause compatibility issues. If an instance of a
sealed class can be exhaustively matched and a new subclass is added to
that sealed class, that is potentially a breaking change for clients who
need to update their exhaustive match statement. Adding the "exhaustive"
or "nonexhaustive" modifier keyword in signature files allows Metalava
to track this between API versions and raise compatibility issues
accordingly.

The "exhaustive" modifier on a sealed class means that instances of the
sealed class can be exhaustively matched (all subclasses are public) and
adding new subclasses is a potentially breaking change. The
"nonexhaustive" modifier on a sealed class means that instances of the
sealed class cannot be exhaustively matched (there is at least one
private/internal/restricted subclass) and adding new subclasses is safe
because clients are forced to have an `else` branch in `when` statements
to make them exhaustive.

Bug: 447143803
Test: N/A
Relnote: N/A
Change-Id: Ie73db64432f9d1327638ada6b2ade4b0656a4c38