Remove @MapInfo
A deprecated API that was replaced by @MapColumn to inform Room which column to map for a non data class type argument.

This change also fixes a missing case of @MapColumn with LongSparseArray and SparseArrayCompat maps as those have hardcoded key types (Long and Int respectively). For those cases a @MapColumn in the declared type itself will be understood to be a mapping column for the key.

Bug: 438041176
Test: Existing
Relnote: "Remove the Room annotation @MapInfo that is deprecated and was replaced by @MapColumn."
Change-Id: I93f8d045b1d1facfa35c281d3e2c01a5d1df91ea