Migrate onPlaced to registerOnLayoutRectChanged()
The dispatching itself is cheaper as we don't need to do the relayout.
Before:
LayoutCoordinatesBenchmark.grandParentMoves_gettingPositionInRoot_onPlaced
timeNs            min 197,955.3,   median 206,796.1,   max 335,626.7
instructions      min  46,091.7,   median  46,164.4,   max  50,705.8
After:
LayoutCoordinatesBenchmark.grandParentMoves_gettingPositionInRoot_onPlaced
timeNs            min 184,598.7,   median 191,400.5,   max 236,710.6
instructions      min  34,976.5,   median  35,096.4,   max  39,231.5
But the main benefit is that the callback will be invoked less often, only when the offset really changed. It was already a contract that we will not invoke the callback when something on the grandparent changes without affecting this node (in comparison to onGloballyPositioned), but we will also not execute the callback each relayout when nothing changes. And we are still keeping the existing contract of callback only being called again if the coordinates were really used.

Test: added new tests (which are passing with the flag disabled, as they test the existing contract)
Relnote: The behavior of Modifier.onPlaced() callback is changed. It will not called on each placement anymore, instead it will only be called when some of the coordinates did change. It will also be called after the layout phase happened instead of being called during the layout, similarly to how Modifier.onGloballyPositioned() works. This change helps to improve the performance. The new feature flag was added to control this behavior - ComposeUiFlags.isNewDispatchingMechanismForOnPlacedEnabled.
Change-Id: Ifcf95d24baea41275ef1e32116f6487095c95d59