Add LocalLocale and LocalLocaleList
Relnote: "Adds LocalLocale and LocalLocaleList as an observable
    source of truth for the current locales."

In preparation for deprecating Locale.current and
LocaleList.current in the follow-up CL, this CL adds the
replacement LocalLocale and LocalLocaleList composition
locals. Both of these are derived proxy composition locals
from LocalProvidableLocaleList, which is an internally providable
composition local, backed by Configuration on Android.

Bug: 240191036
Test: new tests

Change-Id: I67bd55261e36e221662040bf07fa48c93d1c00a3