Protect the frameRateCategoryView variable in AndroidComposeView
Fixes: 460465152

A test is failing that is showing frameRateCategoryView as
uninitialized. The View is initialized in the init block,
so it is likely that the flag controlling the variable is
being manipulated during the test.

This CL makes the variable nullable and checks if it is null. Once
the flag is removed, the variable can be set to a non-null val.

Test: new test

Change-Id: I1b12418679e11209aad048ed1f961a1642643e53