Reduce invalidations due to MANIFEST and CHANGE_INFO env variables
Before this change the value of env variables MANIFEST and CHANGE_INFO
were an input to configuration cache. These values were changing because
the name of them include the build id (e.g. P109429332). We do not
actually care about the name, we only care about the contents of these
files. To achieve that, move to to using ValueSource API. This API only
invalidates the configuration if the returned value changes.

This allows us to undo manifest file copying to OUT_DIR as we no longer
care about the name of the file, only the contents

Test: existing tests pass
      MANIFEST=../../out/dist/manifest_1.xml CHANGE_INFO=CHANGE_INFO_1.json ./gradlew zipTestConfigsWithApks --dry-run -Pandroidx.enableAffectedModuleDetection
      MANIFEST=../../out/dist/manifest_2.xml CHANGE_INFO=CHANGE_INFO_2.json ./gradlew zipTestConfigsWithApks --dry-run -Pandroidx.enableAffectedModuleDetection
      cc hit on the second run
Change-Id: Iaff692202d89482b13fe5cbd9d1d1202a11f9ed3