Unconditionally close capture session during disconnect
SurfaceRequest API documentation stipulates that previous
SurfaceRequests are guaranteed to be detached when a new request is
made. This means whenever we need a new session, all Surfaces should be
disconnected. To do this, we need to close the capture session
unconditionally.

This CL does a thorough cleanup of the behavior customizations in this
area. Previously, we tried to avoid closing the capture session by
default, as is the recommended behavior with performance implications.

Bug: 460221022
Test: JCA hybrid capture switching
Change-Id: I4fdc92ad7ccf3aead76e511c3c9b91a99c9d1f6f