Abstract out SlotStorage implementation
This refactor extracts common classes that the gap buffer and
link table implementations can abstract over. This builds on top of
the previous refactoring commits so that we can have two
implementations of Composer for both implementations of the SlotTable.

There are also a few other generalization changes in this CL to make
dependents of the internal Composer accept an InternalComposer instead
of a GapComposer implementation.

Test: N/A
Relnote: N/A
Bug: 268366116
Change-Id: I08497c6da52547c4fe4ab340e50e782e1ffa36d3