Add XType#isStar().
Currently, there's no great way to determine if a type is a star type
without using the TypeName, but calculating the TypeName can be
expensive. This CL adds a `isStar()` method that returns true if the
given XTYpe represents a star type.

This change saves ~6s on some slow builds.

Test: XTypeTest#isStar()
Change-Id: I4a92ece4b716c1b3b99ea8b6c0ffa81b98f3b120