Introduce RecordingRemoteComposeBuffer
Unlike it's parent class, RecordingRemoteComposeBuffer constructs
a tree of Operations which it reorders to ensure each operation
is in the tightest scope where it's side effects are visible to
all operations that reference it.

Note a number of Operations have been modified to allow them to be
constructed more easily.

Bug: 443939823
Test: New unit tests
Change-Id: Id2edac688e9375b6e25b68e8db49b14b482b0e8c