Deprecate Modifier.onFirstVisible
Relnote: Modifier.onFirstVisible() is deprecated as its behavior is misleading and doesn't always follow the contract claimed by the name. For example, when it is added on an item of LazyColumn, this callback will be called everytime this item became visible after scrolling. It is not what the users of the modifier with this name might have expected. It is recommended to use Modifier.onVisibilityChanged() instead and manually track if the layout was visible already previously based on the requirement of the specific use case.
Test: existing tests
Fixes: 447601783
Change-Id: Ia709566717edbdfe5717bedfa74d1334a78f8964