Allow ImageAnalysis use case with feature groups API
Previously, `SessionConfig` prohibited the simultaneous usage of
`ImageAnalysis` and `GroupableFeature`s, throwing an
`IllegalArgumentException` if both were present. This limited
flexibility for user scenarios where image analysis needs to work in
tandem with specific feature requirements like preview stabilization or
60 FPS.

This change removes that restriction and updates the validation logic.
Specifically, this commit:
- Allows `ImageAnalysis` to be configured alongside feature groups in
  `SessionConfig`.
- Updates `DefaultFeatureGroupResolver` to include `ImageAnalysis` when
  validating stream features.
- Refactors `UseCaseType` to remove unnecessary image format properties
  to enable easier addition of ImageAnalysis support.
- Removes SessionConfig documentation about `ImageAnalysis` not being
  supported with feature groups.
- Updates unit tests and the integration test app to verify support for
  `ImageAnalysis` in feature combinations.

RelNote: "Added support for using both the ImageAnalysis and the feature
groups API in SessionConfig."

Bug: 406371720
Test: FeatureCombinationDeviceTest, FeatureGroupQueryBindAlignmentTest, manual app check
Change-Id: I87833dab66b27899d9c8740caa85c5427a45bcc1