Update configured session even if CaptureSessionState was disconnected
During shutdown, we could be in a state where we have attempted a
capture session, but we haven't received the configured session yet.
If we also need to close the capture session (to disconnect the
Surfaces), we need to close the capture session when the configured
session comes back.

Here the issue is CaptureSessionState doesn't update the configured
capture session variable when it is shut down. This CL makes sure we
update the variable, so that the session can be closed and finalized
during CaptureSessionState.disconnect().

Bug: 461964604
Test: Being tested
Change-Id: I7e4bc855e2778e9f81a3552c31e7723e05204136