Update Metalava command to use --api-version-for-sources
This CL:
* Removes metalava arg for --current-version
* Add metalava arg for --api-version-for-sources

Context:
b/447588979
The current version argument is being deprecated in favour of a new more
descriptive one. This simplifies the amount of arguments required to set
api versions in the api history. It also allows Metalava to explicitly
set the api version for the upcoming releases (In the past, it would be
assumed that the next release would be current+1 which is no longer true
now we have minor SDK release e.g. 36.1)

Bug: 454050901
Test: m
Change-Id: I712b0388982b6ece7659fbb724b9abd89f351f78