Be more explicit w/ WorkerWrapper Resolutions
As preparation for debug APIs, this CL refactors the internal Resolution
object to match the semantic meaning of the result more instead of just
saying how it should be handled.

e.g. being "Stopped" and a WorkerWrapper pre-condition failure may both
need to reset worker state but are semantically different triggers

This will be needed in order to ensure certain debug hooks only trigger
on the appropriate event e.g. stop. There are no functional changes in
this CL, just refactoring the names.

To summarize the changes:

ResetWorkerStatus => Stopped or WorkerWrapperFailure(recoverable=true)
Failed => Failed (due to app) or WorkerWrapperFailure(recoverable=false)

Bug: 448485695
Test: WorkerWrapperTest
Change-Id: I3a2944038d0d407c97c4aa5ee64d029fee06840f