Remove ArCoreManager.running
Remove the internal variable within ArCoreManager for keeping track of
whether the underlhing ARCore 1.x session is in the resumed state. In
order to do this, ensure that update is not called on the underlying
session while it is paused by swapping the order of the cancellation of
the update job with pausing all of the runtimes of the JXR Session in
Session.pause().

Bug: 411154789
Test: On Pixel 9 Pro using whitebox mobile app and unit tests
Change-Id: Iff79928573ef79f37f84b3d7905519fae13cd2a8