Only acquire position for sharedElements if SharedTransitionLayout is attached
This CL adds an additional check in the detaching process of a
sharedElement/sharedBounds modifier, so that we only calculate
the last position before detaching if the SharedTransitionLayout
is still attached.

This additional check is needed to accommodate the use case where
sharedElement modifier and SharedTransitionLayout are removed
from the tree within the same frame, while the content that was
previously in SharedTransitionLayout got moved out using MovableContent.

Bug: 462433859
Test: new test added
Change-Id: I2a03554b3528abf628787091d3a7bede6c01a292