Add InputTextSuggestionState and TextCompositionRange
InputTextSuggestionState stores a snapshot of the input state when
text replacement suggestions are present for the inputted text.
Transliteration languages, such as Chinese, Japanese, and Korean,
often have suggestion popups when inputting text. The goal is to
provide additional information for accessibility event handling.

TextCompositionRange provides the composition range of the current
text. If null, there is no transliteration composition.

Relnote: "Add a `InputTextSuggestionState` property to help identify
the different states a transliteration input can be in when text
replacement suggestions are available. Also introduces
`TextCompositionRange` to track the range of the current
transliteration text composition range; a null value indicates no
composition in place."

Bug: 476472318
Test: TextFieldStateTest

Change-Id: I7a67fc9413c1b856af1c8a41ff25cbc0648c5e97