Add Implementation of Api onFirstContentLoad
This Cl ads the Implementation for onFirstContentLoad api, where the callback is provided to inform the users that First time content of the document has been loaded, this is fired once new document loads to maintain a consistent contract and to not expose any internal optimizations.
This is achieved by tracking when the bitamp for the visible page has been fetched and on subsequent onDraw firing the callback on the post block, and the event is deduplicated using shouldNotifyFirstContentLoad boolean which resets on either view recreation or resetting the document.

Test: ./gradlew :pdf:pdf-viewer:test --tests androidx.pdf.view.FirstContentLoadTest
Bug: 461565451
Change-Id: Icc34e0dd712a5d1a73b67412b155752730d19c13