Fix remoteDrawAnchoredText in RemoteCanvas.
Replace calls to getFloatIdForCreationState with direct access to the .id property for RemoteFloat instances in remoteDrawAnchoredText functions. This simplifies the code as alpha and textSize are always RemoteFloat since aosp/3852562. It also fixes the code as alpha and textSize could be of type RemoteFloatExpression, which overrides the implementation of RemoteFloat.id.

Test: screenshot tests updated.
Change-Id: I46bde3a722091a9804790048a474bde0ee0aee6a