Removing RememberComposeView from SpatialDialog implementation and replacing it with a ComposeView in a RememberObserver.
It has been our goal to remove all usage of RememberComposeView from base implementations of all Spatial Composables after API +2.

Its intended purpose is to create a ComposeView that is applicable to the local context in order to handle propagating the compositional context and could
be saved in a remember state. This design of rememberComposeView() was functional in developing other Spatial UI components, however it is apparent now that
its definition does not follow best practices of the Compose runtime and has led to limitations being next to difficult or impossible to solve.

By utilizing a ComposeView within a RememberObserver for our base Orbiter implementation we resolve the following issues:

1) We should neither create nor manipulate a View by returning it from the Composition hierarchy
2) 2D Compose UI content should be created as a subcomposition in the XR content
3) The view's composition creation and disposal logic must not be decoupled from its presentation

More information provided in the following doc:
go/compose-xr:deprecating-remembercomposeview

Test: layout logic for SpatialDialog confirmed with TestApp

Bug: 415130218
Change-Id: I55ffc83e44bd22b1168aa83c9276be31d1abd0be