Fix flickering when spawning AnchorEntity
Currently, AnchorEntityActivity initializes GltfModelEntity with ActivitySpace as its default parent. This triggers immediate rendering on the scene graph, causing visible flickering when the entity is subsequently reparented to an AnchorEntity.

This patch initializes the GltfModelEntity parent as null to prevent immediate rendering. The entity is reparented and enabled only when it is ready to be displayed.

Bug: b/450028067
Test: On the Moohan device
Change-Id: I441307f1d0a78b8a80119e7692217f5ead52f36c