Fix initial ContentInViewNode viewportSize
ContentInViewNode sets its initial viewportSize to `IntSize.Zero` but then uses this value as an initialization flag. When the node itself actually shrinks down to Zero size, it is inferred as non-placed node. The fix is to start with an Unspecified value that can also be inferred as `IntSize.Zero` from places that do not care about placement.

Test: TextFieldScrollTest
Fix: 440964236
Change-Id: I653a6367e2d9dd3022dd3d83f8d59ca4db25dc72