Add ATOMIC_MOVE copy option
When we replace the existing datastore file with the temporary file's
contents, we use the REPLACE_EXISTING copy option. We should also use
the ATOMIC_MOVE option to ensure that the move is performed as an
atomic operation.

Also refactored the atomicMoveTo function so that it no longer returns
a boolean value to indicate success, but instead throws the underlying
exception that we then include as cause for the IOException that we
throw. This will help with debugging future issues.

This is a speculative fix for b/203087070, but I think we should do
this anyway.

Bug: 203087070
Test: Verified that existing tests run.
Change-Id: I94f89debfd53d8c1d33e97a9725ed9884a1a2da2