Reference App: enable heads up alert call-style notifications
Previously, incoming call-style notifications were not displaying
in the foreground and going silently into the notifications tray
with a small alert in the bar.

This was due to the NotificationCompat.setSilent(false) value which
I thought was only for sound.

Toggling NotificationCompat.setSilent(true) will display notifications
as heads up / foreground alerts

Test: manual
Bug: 465212048
Change-Id: I7b5e297e16ccfc6825586dccf62a4133ca2f9aca