Raise minSdk to 29 and clean up code
This commit increases the `minSdk` for the `remote-creation-compose` library from 26 to 29. RecordingCanvas uses BlendMode so it can't be lower.

This change allows for the removal of compatibility code and API level checks that were previously required for older Android versions, particularly related to gradients and blend modes.

Additionally, this commit includes several code cleanup improvements:
*   Replaces `HashMap` with `MutableObjectIntMap` and `MutableIntObjectMap` for better performance with primitive types.
*   Addresses various lint warnings by adding `@Suppress` annotations for `PrimitiveInCollection` and `ListIterator`, and by fixing an unnecessary lambda creation.
*   Changes `mutableStateOf<Long>` to `mutableLongStateOf` to avoid autoboxing.

Test: Existing tests pass
Change-Id: I6c57592defb90356ca32a6c14bd9239b8b4f835e