Remove unnecessary role from volume slider content
As per View#setContentDescription, the content description (which is
defined in this string value) should not include the role of the view.
The role is detected separately by the accessibility tools.

I've intentionally left the role in the comment next to the string for
consistency with other accessibility strings in the same file.

Recordings for comparing before/after this change are available in the
linked ticket.

Bug: b/445871174
Test: Manually using the media routing demo app with talkback.
Change-Id: I3c2ebef43c160269d65495dfffefaf494c748ab4