Skip running use case update when no use cases attached
Currently, UseCaseManager#refreshRunningUseCases proceeds
to calculate the set of running UseCases and attempts to
update the repeating request even when the list of attached
UseCases is empty.

When attachedUseCases is empty, it indicates that the camera
session is either not established or is in the process of
being torn down (e.g., during a detach operation where the
count goes from 1 to 0). In this state, triggering an update
to the repeating request is unnecessary and redundant.

This CL adds an early return check to refreshRunningUseCases.
If no UseCases are attached, we skip the update logic.

Bug: 448593362
Test: UseCaseManagerTest
Change-Id: I7cda394652bcef7716396628edfabb427485fc3c