Update metalava source paths and dependency classpath for KMP projects
Previously, the source paths and classpath for a `MultiplatformCompilationInputs` were aggregated from the source paths and classpath of all source sets included in the inputs, which is currently just the source sets involved in the jvm or android compilation. However, the followup change will start including all source sets in the MultiplatformCompilationInuts. To preserve the existing behavior for kzip generation, this sets the source paths and classpath to specifically use the values from the main jvm or android compilation.

Bug: 407737495
Test: ./gradlew checkApi
Change-Id: I87c1e38f0f96f113f366a5a30a9764516fce21cc