Fix movableContentOf SelectionContainer crash
Removing a movableContent from composition when there is a text selection in it causes a crash due to an internal position and bounds calculation. This calculation does not take the attached status of layoutCoordinates into account.

The fix is to skip the calculation when either the text or container coordinates are unattached.

Fix: 444482508
Test: SelectionContainerTest
Change-Id: I9f5408fbfc56be0de9632d979b4c3e89a9c59139