Use correct kotlin platforms in metalava KMP information
The current KMP information generated for metalava lists JVM as the only platform for all source sets. To include non-jvm/android targets, this needs to be updated to list native/wasm/js when required. For any source set, the list of platforms should include all platforms that the source set is involved in a compilation for.

Bug: 324624680
Bug: 407737495
Test: Tested running metalava with pending changes to enable multiplatform codebase creation and verified that there were no unresolved types. Updated ProjectXmlTest.

Change-Id: Ia5b90200c19144e21ad642324c39327509e7583d