Handle ctrl-click via classifySelectionHotSpot
Prior to this commit, mapping "control-click" to
SELECTION_HOTSPOT_INSIDE_TOGGLE_MULTI was the responsibility of the r-s
(recyclerview-selection) library, not the r-s callback implementor - the
app (not library) code that overrides classifySelectionHotspot.

After this commit, it's now the responsibility of the callback
implementor, giving r-s users more control about handling control-click.

classifySelectionHotspot is a relatively recent API addition, and an
optionally-overridable method, so the number of callback implementors
out there in the world is expected to be in the single digits. One
example is Android's packages/apps/DocumentsUI.

If an app doesn't override the method, there should be no behavior
change for mouse clicks (regardless of whether or not control is down).

Note that this commit does not modify TouchInputHandler.java. Strictly
speaking, this means that selection behavior might change, when handling
a finger-tap (not a mouse-click) that *also has a control-key modifier
pressed*. But that sounds like a really rare situation.

Bug: 429096800
Test: gradlew :recyclerview:recyclerview-selection:connectedAndroidTest
Relnote: """Ctrl-click handling is now the responsibility of `ItemDetailsLookup.ItemDetails#classifySelectionHotspot`. Apps that want to support ctrl-click multi-selection must do so by returning `ItemDetails.SELECTION_HOTSPOT_INSIDE_TOGGLE_MULTI` when `MotionEvents.isCtrlKeyPressed(e)` is `true`. Apps using the default implementation are unaffected and require no changes."""
Change-Id: Ie0a2c163b89c6519ccf6eedcc5a1fa34c4c091cb