Add web targets to SQLite driver APIs in androidx.sqlite
To support web targets suspend / async versions of key APIs are added to the interfaces, openAsync(), prepareAsync() and stepAsync(). These APIs are added in common code that supports all platforms, while the sync versions are only supported on non-web. The interfaces are refactored to be actual / expect declarations to be able to evolve the API such that existing users that do not target web are not affected.

Once a common source set using driver APIs targets web, then the blocking APIs will become unavailable and the user must use the suspending ones. If the common source set does not target web, then usage suspending APIs will be hidden, forcing users to continue using the blocking ones. This last bit is done by adding the suspend versions in the nonWeb source set as @Deprecated(level = HIDDEN).

To facilitate common code callers that target both web and non-web then suspending extension functions are available that are expect / actual and will call the corresponding API based on the platform.

Bug: 336758416
Bug: 438041176
Test: ./sqlite/scripts/runConformanceTest.sh
Relnote: Add js and wasmJs targets to androidx.sqlite along with SQLiteDriver.openAsync(), SQLiteConnection.prepareAsync() and SQLiteStatement.stepAsync() APIs to enabling creating async SQLite Drivers for the web.
Change-Id: Id54702011351d9b42ffa5799e0da0a743e538453