Sync animation state in Scenecore to avoid crash on Pause.
When an animation plays and reaches the end of its animation,
impress updates the internal state to STOPPED,
but we never updated it in Scenecore.

This bug allowed scenecore to send a Pause request to Impress when
Impress an animation was already stopped. This is not allowed so impress
would throw an exception and crash the app.

Bug: 467045293
Test: Tested in the test apps and unit tests
Change-Id: I1e01a65412901a916d7aba0eaa780dcd9317500a