Lint against ignored runTest results
On Kotlin/JS, runTest returns a promise. If the result is not returned
by a `@Test` function (or manually awaited), the test lambda will
never execute. Effectively, this means that tests that don't return
the result of runTest can never fail on JS and WASM targets.

This CL adds an internal lint inspection to enforce using the return
result of runTest. Follow-up CLs will correct the errors found.

Test: KotlinRunTestReturnResultDetectorTest
Fixes: b/444012850
Relnote: N/A
Change-Id: Ia8549ff5a1fa90ee09ad9c497502abd32caf83d6