Update baseline profile files
We added baseline profiles in aosp/3867894, but some
of the baseline profiles were added to modules that
don't produce an AAR. This CL fixes that problem by
adding the baseline profiles to the modules that use
the dependencies instead.

As part of the baseline profile generation, we use
the :development:splitBaselineProfiles utility which
splits the baseline profiles into multiple modules.
The baseline profile files now include the comments
created by the utility, to make it easier to update
the baseline profiles in the future.

Bug: 366076962
Bug: 469127532
Test: Verified that the AAR contains the baseline profiles
Change-Id: I9256b0d3ad0029a40b29d9de4efc6844c4fd2e8e