Extend LicenseDownloader timeout to 1 minute
The default OkHttp timeout seems to be 10s, and shouldn't have changed
recently, so I'm not sure why this is needed now (since ~October).

Possibly either something changed on the license proxy, or on certain
hosts we read licenses from.

When a license fails to download (including all retries failing) it
seems to cause corrupted/invalid output from `importMaven` without
actually causing the overall script execution to fail (so subsequent
operations may continue) - see b/454286734 for details. My repro
attempts since aosp/3877343 saw timeouts, but they were resolved on a
retry. So it's possible this behaviour has changed. However, assuming
the corrupted output/successful execution is possible, the penalty for
a failed license download is high, so we deliberately choose a high
timeout to reduce the risk at the cost of possibly slower execution of
the `importMaven` task.

Bug: 454286734
Test: ./development/importMaven/importMaven.sh org.robolectric:robolectric:4.13
Change-Id: If7153999b8e663d7430355541179db68b4c56b04