Return CapturedDocument from captureRemoteDocument
The `captureRemoteDocument` function now returns a `CapturedDocument` object instead of a `ByteArray`.

This new `CapturedDocument` class encapsulates both the serialized document bytes and any associated `PendingIntent`s that were captured. This simplifies the function's signature by removing the need to pass in a `WriterEvents` callback, as it is now managed internally.

Relnote: Return CapturedDocument from captureSingleRemoteDocument
Test: Existing
Bug: 467532762
Change-Id: I5a2832bcd1eb4b54e82a413452d1146565925b81