Stop reading itemWidth in composition of NavigationItem
The MappedInteractionSource requires an offset to shift interactions that is calculated based on `itemWidth`. This results in a composition loop that can be avoiding by moving `itemWidth` state read a lambda which this change does.

The offset is only used on press interaction and is not required until press actually occurs.

Test: Existing NavigationItem tests
Change-Id: Ieb5a343177700e7b82884680ba50a7eeded737e6